fixed the sparql data flow for batched lookups

+++ jena-sparql-api-utils/src/main/java/org/aksw/jena_sparql_api/utils/QueryUtils.java
@@ -190,10 +190,10 @@ public class QueryUtils {
 
         Long newMax = (parent.hasUpperBound()
             ? child.hasUpperBound()
-                ? Math.min(parent.upperEndpoint(), child.upperEndpoint())
+                ? (Long)Math.min(parent.upperEndpoint(), child.upperEndpoint())
                 : parent.upperEndpoint()
             : child.hasUpperBound()
-                ? child.upperEndpoint()
+                ? (Long)child.upperEndpoint()
                 : null);
 
         Range<Long> result = newMax == null
